鄭家霖.修改:新增Interface (#22)
