+++ pom.xml
@@ -73,7 +73,10 @@
 	<build>
 		<resources>
 			<resource>
-				<directory>src/main/java</directory>
+				<directory>src/main/java</directory>	
+			</resource>
+			<resource>
+				<directory>src/main/resources</directory>	
 			</resource>
 		</resources>
 		<plugins>
