+++ src/main/java/com/booking/dao/attraction/AttractionDao.java
@@ -1,38 +1,18 @@
 package com.booking.dao.attraction;
 
-import java.util.ArrayList;
 import java.util.List;
-
-import org.hibernate.Session;
-import org.hibernate.query.Query;
 
 import com.booking.bean.attraction.Attraction;
 import com.booking.utils.util.DaoResult;
 
-import jakarta.persistence.criteria.CriteriaBuilder;
-import jakarta.persistence.criteria.CriteriaQuery;
-import jakarta.persistence.criteria.Predicate;
-import jakarta.persistence.criteria.Root;
-
-public class AttractionDao {
-
-	private Session session;
-
-	public AttractionDao(Session session) {
-		this.session = session;
-	}
+public interface AttractionDao {
 
 	/**
 	 * 獲取所有景點
 	 * 
 	 * @return
 	 */
-	public DaoResult<List<Attraction>> getAttractionAll() {
-		String hql = "From Attraction";
-		Query<Attraction> query = session.createQuery(hql, Attraction.class);
-		List<Attraction> attractions = query.getResultList();
-		return DaoResult.create(attractions).setSuccess(attractions != null);
-	}
+	DaoResult<List<Attraction>> getAttractionAll();
 
 	/**
 	 * 模糊查詢
@@ -40,58 +20,7 @@ public class AttractionDao {
 	 * @param attraction
 	 * @return
 	 */
-	public DaoResult<List<Attraction>> dynamicQuery(Attraction attraction) {
-		String attractionName = attraction.getAttractionName();
-		String attractionCity = attraction.getAttractionCity();
-		String address = attraction.getAddress();
-		String openingHours = attraction.getOpeningHour();
-		String attractionType = attraction.getAttractionType();
-		String attractionDescription = attraction.getAttractionDescription();
-
-		// 透過session創建CriteriaBuilder
-		CriteriaBuilder cb = session.getCriteriaBuilder();
-		// 創建CriteriaQuery查詢
-		CriteriaQuery<Attraction> cq = cb.createQuery(Attraction.class);
-		Root<Attraction> root = cq.from(Attraction.class);
-
-		// where數據篩選
-		List<Predicate> predicates = new ArrayList<>();
-
-		if (attractionName != null) {
-			predicates.add(cb.like(root.get("attractionName"), "%" + attractionName + "%"));
-		}
-
-		if (attractionCity != null) {
-			predicates.add(cb.like(root.get("attractionCity"), "%" + attractionCity + "%"));
-		}
-
-		if (address != null) {
-			predicates.add(cb.like(root.get("address"), "%" + address + "%"));
-		}
-
-		if (openingHours != null) {
-			predicates.add(cb.like(root.get("openingHours"), "%" + openingHours + "%"));
-		}
-
-		if (attractionType != null) {
-			predicates.add(cb.like(root.get("attractionType"), "%" + attractionType + "%"));
-		}
-
-		if (attractionDescription != null) {
-			predicates.add(cb.like(root.get("attractionDescription"), "%" + attractionDescription + "%"));
-		}
-
-		// 篩選條件
-		cq.select(root).where(predicates.toArray(new Predicate[0]));
-
-		// 創建session查詢
-		Query<Attraction> query = session.createQuery(cq);
-
-		// 透過session查詢獲得結果
-		List<Attraction> attractions = query.getResultList();
-
-		return DaoResult.create(attractions).setSuccess(attractions != null);
-	}
+	DaoResult<List<Attraction>> dynamicQuery(Attraction attraction);
 
 	/**
 	 * 依id獲取景點
@@ -99,12 +28,7 @@ public class AttractionDao {
 	 * @param attractionId
 	 * @return
 	 */
-	public DaoResult<Attraction> getAttractionById(Integer attractionId) {
-		String hql = "FROM Attraction WHERE attractionId = :attractionId";
-		Attraction attraction = session.createQuery(hql, Attraction.class).setParameter("attractionId", attractionId)
-				.getSingleResult();
-		return DaoResult.create(attraction).setSuccess(attraction != null);
-	}
+	DaoResult<Attraction> getAttractionById(Integer attractionId);
 
 	/**
 	 * 新增景點
@@ -112,34 +36,20 @@ public class AttractionDao {
 	 * @param attraction
 	 * @return
 	 */
-	public DaoResult<?> addAttraction(Attraction attraction) {
-		session.persist(attraction);
-		Integer attractionId = attraction.getAttractionId();
-		return DaoResult.create().setGeneratedId(attractionId).setSuccess(attractionId != null);
-	}
+	DaoResult<?> addAttraction(Attraction attraction);
 
 	/**
 	 * 依id刪除景點
 	 * 
 	 * @param attractionId
 	 */
-	public DaoResult<?> removeAddractionById(Integer attractionId) {
-		Attraction attraction = session.get(Attraction.class, attractionId);
-		if (attraction != null) {
-			session.remove(attraction);
-			return DaoResult.create().setSuccess(true);
-		}
-		return DaoResult.create().setSuccess(false);
-	}
+	DaoResult<?> removeAddractionById(Integer attractionId);
 
 	/**
 	 * 更新景點
 	 * 
 	 * @param attraction
 	 */
-	public DaoResult<?> updateAttraction(Attraction attraction) {
-		Attraction updatedAttraction = session.merge(attraction);
-		return DaoResult.create().setSuccess(updatedAttraction != null);
-	}
+	DaoResult<?> updateAttraction(Attraction attraction);
 
 }
